"Updated dependencies, refactored AreaChartComponent and BookDetails components, and made changes to deleteBook and updatePermissions functions."

+++ app/dashboard/_actions/deleteBook.ts
@@ -6,21 +6,35 @@ export async function DeleteBook(id: string) {
   if (!id) return { success: false, message: 'Book cannot be identified' };
 
   try {
+    const existingBook = await prisma.book.findUnique({
+      where: { id },
+    });
+
+    if (!existingBook) {
+      return { success: false, message: 'Book not found' };
+    }
+
     await prisma.readingProgress.deleteMany({
       where: { bookId: id },
     });
 
-    const deletedBook = await prisma.book.delete({
-      where: {
-        id,
-      },
-    });
+    await prisma.bookPermission.deleteMany({
+      where: {bookId: id}
+    })
 
-    if (!deletedBook) return { success: false, message: 'Book Not Found!' };
+    const deletedBook = await prisma.book.delete({
+      where: { id },
+    });
 
     return { success: true, data: deletedBook };
   } catch (error) {
-    console.error('Error deleting book', error);
-    return { success: false, message: 'unable to delete book.' };
+    console.error('Error deleting book:', error);
+    return {
+      success: false,
+      message:
+        error instanceof Error
+          ? `Unable to delete book: ${error.message}`
+          : 'Unable to delete book',
+    };
   }
 }
